requirements.txt
# Core dependencies
pydantic>=2.0
click>=8.0
rich>=13.0
pyyaml>=6.0
jinja2>=3.0
croniter>=1.0

# Development dependencies
pytest>=7.0
pytest-cov>=4.0
mypy>=1.0
ruff>=0.1.0
